lwt: if connect() blocks, timeout after 1s and ECONNREFUSED
If the server isn't listening then connect appears to block forever,
even if the server subsequently starts up. This patch simulates an
ECONNREFUSED which should cause the client to try again with a fresh
socket.

Note we also wake up the blocked background thread with an explicit close
and avoid double-closes by using a record with a mutable file_descr option.

Signed-off-by: David Scott <dave@recoil.org>