Flow_lwt_hvsock: on unexpected read/write errors: return `Eof
Up until now we've been assuming a Unix convention of

- read returns 0 on Eof, other errors mean programming error
- write fails with EPIPE on Eof, other errors mean programming error

However Windows AF_HYPERV is not guaranteed to stick to the same narrow
interface: other errors which mean Eof are believed to be possible.

This patch logs such unexpected errors for analysis and returns `Eof
which should trigger a cleanup.

Signed-off-by: David Scott <dave.scott@docker.com>