Use Cstruct.t for packet buffers rather than Bytes.t
Previously we used Bytes.t here because we needed a blocking call
to read and write buffers to a SOCKET, and the standard lib only has
Unix.read and Unix.write. This meant we had to copy back and forth
between Cstruct.t (as used everywhere else) and Bytes.t (used here).

This patch switches us to use Cstruct.t, and implements blocking SOCKET
`recv` and `send`  in C.

Signed-off-by: David Scott <dave.scott@docker.com>