Fix `write` in Flow_lwt_hvsock_shutdown
When we compute how many bytes to send (the minimum of the maximum
message size and the number of bytes in our hand) and write this in
the header, we MUST write exactly this number of bytes. Previously
we would write as many bytes as we had available.

Previously if we had 5000 bytes to write, we would write 4096 in the
header (since `min 4096 5000 = 4096`) and then write 5000 anyway,
corrupting the following header.

Signed-off-by: David Scott <dave.scott@docker.com>