Properly guard boolean signalling variables with mutexes
Previously this code was written in the "lwt style" i.e. where it is
safe to assume no other threads are interleaving with the current
function, apart from at explicit bind points. We are actually using
pthreads here! This means that arbitrary interleavings are possible so
we must take care to lock mutexes around the variables which we are
sharing between threads for signalling purposes.

Signed-off-by: David Scott <dave@recoil.org>